# gen/gen.models.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

// ErrorResponce defines model for ErrorResponce.
type ErrorResponce struct {
	ErrMsg *string `json:"err_msg,omitempty"`
}

// Schedule defines model for Schedule.
type Schedule struct {
	DurationMedication *int    `json:"duration_medication,omitempty"`
	Id                 *int    `json:"id,omitempty"`
	MedicationPerDay   *int    `json:"medication_per_day,omitempty"`
	NameMedication     *string `json:"name_medication,omitempty"`
	UserId             *int    `json:"user_id,omitempty"`
}

// ScheduleTo defines model for ScheduleTo.
type ScheduleTo struct {
	Id               *int    `json:"id,omitempty"`
	MedicationPerDay *int    `json:"medication_per_day,omitempty"`
	NameMedication   *string `json:"name_medication,omitempty"`

	// Schedule Array of times for taking the pill in coming period
	Schedule *[]string `json:"schedule,omitempty"`
}

// CreateScheduleJSONRequestBody defines body for CreateSchedule for application/json ContentType.
type CreateScheduleJSONRequestBody = Schedule